minor modifications for weight decay etc

+++ learner/brain_FNN.py
@@ -273,9 +273,9 @@ class Brain_FNN:
 #             else:
 #                 loss_AE_jac, J_e, J_d, idx_trunc = self.SAE.jacobian_norm_trunc_gpu(z_gt_tr_norm, x, self.trunc_period)
             if  ((self.lambda_jac == 0 and self.lambda_dx == 0) and self.lambda_dz == 0): 
-                loss_AE_jac = torch.tensor(0)
-                loss_dx = torch.tensor(0)
-                loss_dz = torch.tensor(0)
+                loss_AE_jac = torch.tensor(0, dtype=torch.float64)
+                loss_dx = torch.tensor(0, dtype=torch.float64)
+                loss_dz = torch.tensor(0, dtype=torch.float64)
                 
             else:
 
@@ -341,9 +341,9 @@ class Brain_FNN:
                 loss_AE_recon_test = torch.mean((z_sae_tt_norm - z_gt_tt_norm) ** 2)
                 
                 if  ((self.lambda_jac == 0 and self.lambda_dx == 0) and self.lambda_dz == 0): 
-                    loss_AE_jac_test = torch.tensor(0)
-                    loss_dx_test = torch.tensor(0)
-                    loss_dz_test = torch.tensor(0)
+                    loss_AE_jac_test = torch.tensor(0, dtype=torch.float64)
+                    loss_dx_test = torch.tensor(0, dtype=torch.float64)
+                    loss_dz_test = torch.tensor(0, dtype=torch.float64)
                 else:
 
                     J_ed, J_e, J_d, idx_trunc = self.SAE.jacobian_norm_trunc_wo_jac_loss(z_gt_tt_norm, x_tt, self.trunc_period)
